requirements.txt
task
Django==2.0
djangorestframework==3.7.7
psycopg2==2.7.3.2
